backend/requirements.txt
# ── Web framework ─────────────────────────────────────────────────────────────
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.6.0
flask-socketio==5.3.6
flask-mail==0.10.0
flask-cors==4.0.1
flask-migrate==4.0.7
werkzeug==3.0.3

# ── WSGI server ───────────────────────────────────────────────────────────────
gunicorn==22.0.0
eventlet==0.36.1

# ── Database ──────────────────────────────────────────────────────────────────
psycopg2-binary==2.9.9   # PostgreSQL driver (production)
# SQLite ships with Python — no extra package needed for local dev

# ── Cache / Queue ─────────────────────────────────────────────────────────────
redis==5.0.7
celery==5.4.0

# ── Data processing ───────────────────────────────────────────────────────────
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
openpyxl==3.1.2           # Excel .xlsx read/write
python-docx==1.1.2        # .docx ingestion

# ── NLP ───────────────────────────────────────────────────────────────────────
spacy==3.7.4
# After pip install, run: python -m spacy download en_core_web_sm

# ── OCR / AI ─────────────────────────────────────────────────────────────────
google-generativeai==0.7.2   # Gemini Vision API (replaces Tesseract + pdf2image)
pypdf==4.2.0                 # Pure-Python PDF text extraction (no poppler needed)
Pillow==10.3.0               # Image handling

# ── File storage ──────────────────────────────────────────────────────────────
cloudinary==1.40.0

# ── Notifications ─────────────────────────────────────────────────────────────
twilio==9.2.3

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
requests==2.32.3

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==8.2.2
pytest-flask==1.3.0
